wip but adds PageNotFound page and new firebase realtime database stuff
