+++ src/App.js
@@ -6,6 +6,7 @@ import Currency from './components/pages/Currency';
 import Landing from './components/pages/Landing';
 import Landmarks from './components/pages/Landmarks';
 import Translate from './components/pages/Translate';
+import PageNotFound from './components/pages/PageNotFound';
 
 import { ChakraProvider } from "@chakra-ui/react";
 
@@ -20,6 +21,7 @@ function App() {
         <Route path='translate' element={<Translate />} />
         <Route path='currency' element={<Currency />} />
         <Route path='landmarks' element={<Landmarks />} />
+        <Route path='*' element={<PageNotFound />} />
       </Routes>
       <Footer />
       </ChakraProvider>
