+++ src/components/pages/Landmarks.js
@@ -4,7 +4,6 @@ function Landmarks() {
   return (
     <main>
       <h1>Landmarks</h1>
-      <ShareASuggestion />
     </main>
   );
 }
